Remove swap_matrix.py and add swap_matrix_visualization.py for G_t matrix comparison and visualization
